+++ apps/docs/content/_meta.js
@@ -10,12 +10,16 @@ export default {
   "03-concepts": "Concepts",
   "04-reference": "Reference",
   "05-troubleshooting": "Troubleshooting",
+  "--developer": {
+    type: "separator",
+    title: "Developer docs",
+  },
+  "06-development": "Development",
   "--contributing": {
     type: "separator",
     title: "Contributing & Policies",
   },
   "07-contributing": "Contributing",
   security: "Security",
-  "06-development": "Development",
   about: "About",
 };